use random_features in randomize_graph
