+++ openproblems/tasks/_batch_integration/batch_integration_feature/methods/baseline.py
@@ -1,7 +1,6 @@
 from .....tools.decorators import method
 from .....tools.utils import check_version
-
-import numpy as np
+from ...batch_integration_graph.methods.baseline import _randomize_features
 
 
 @method(
@@ -15,18 +14,6 @@ import numpy as np
 def no_integration(adata, test=False):
     adata.uns["method_code_version"] = check_version("openproblems")
     return adata
-
-
-def _randomize_features(X, partition=None):
-    X_out = X.copy()
-    if partition is None:
-        partition = np.full(X.shape[0], 0)
-    else:
-        partition = np.asarray(partition)
-    for partition_name in np.unique(partition):
-        partition_idx = np.argwhere(partition == partition_name).flatten()
-        X_out[partition_idx] = X[np.random.permutation(partition_idx)]
-    return X_out
 
 
 @method(
